requirements.txt
streamline
pillow
time
random